Refactor chat agent integration
