+++ internal/agentapp/ids.go
@@ -1,0 +1,14 @@
+package agentapp
+
+import (
+	"crypto/rand"
+	"encoding/hex"
+)
+
+func newID(prefix string) string {
+	var bytes [16]byte
+	if _, err := rand.Read(bytes[:]); err != nil {
+		panic(err)
+	}
+	return prefix + "_" + hex.EncodeToString(bytes[:])
+}
